20.03.23

+++ 삼성 A형 기출/[17070] 파이프 옮기기 1/[17070] 파이프 옮기기 1/main.cpp
@@ -28,7 +28,7 @@ Pipe sero(Pipe);
 Pipe daegak(Pipe);
 
 int N;
-int cnt = 0;
+long long int cnt = 0;
 Pipe pipe;
 
 int main(int argc, const char * argv[]) {
